installer.py налаженъ

+++ installer.py
@@ -34,9 +34,10 @@ time.sleep(3)
 print("The installation of the custom components is complete. Move on to system configuration...") 
 os.system('sudo localectl set-x11-keymap --no-convert us,ru pc105 "" grp:alt_shift_toggle')
 os.system('sudo localectl set-locale ru_RU.UTF-8')
-os.system('sudo echo "FONT=cyr-sun16" >> /etc/vconsole.conf') 
-os.system('sudo systemctl restart vconsole')
+
+time.sleep(2) 
 
 print("Installation complete")
-print("now rebooting pc")
-os.system('sudo reboot') 
+rebootnoyes = input('Reboot PC? [y/n]:')
+if rebootnoyes == "y":
+    os.system('sudo reboot')
